Implemented comments from Dhar
