+++ src/app/core/models/advance-approver.model.ts
@@ -1,0 +1,14 @@
+export interface AdvanceApprover {
+  added_by: string;
+  advance_request_id: string;
+  approver_email: string;
+  approver_id: string;
+  approver_name: string;
+  approver_org_id: string;
+  comment: string;
+  created_at: Date;
+  disabled_by: string;
+  id: number;
+  state: string;
+  updated_at: Date;
+}
