+++ src/app/fyle/my-view-advance-request/my-view-advance-request.page.html
@@ -62,7 +62,7 @@
           <div class="view-advance-request--element">
             <app-fy-statistic
               [icon]="'../../../../assets/svg/no-attachment.svg'"
-              [type]="'string'"
+              [type]="statisticType.string"
               [label]="'Receipt'"
               [value]="'Not Attached'"
             ></app-fy-statistic>
@@ -74,7 +74,7 @@
         <div class="view-advance-request--element">
           <app-fy-statistic
             [icon]="'../../../../assets/svg/advances.svg'"
-            [type]="'string'"
+            [type]="statisticType.string"
             [label]="'Request ID'"
             [value]="advanceRequest.areq_advance_request_number|| '-'"
           ></app-fy-statistic>
@@ -88,7 +88,7 @@
               <div class="view-advance-request--element">
                 <app-fy-statistic
                   [icon]="'../../../../assets/svg/calendar.svg'"
-                  [type]="'date'"
+                  [type]="statisticType.date"
                   [label]="'Created On'"
                   [value]="advanceRequest.areq_created_at | date: 'MMM dd, yyyy'"
                 ></app-fy-statistic>
@@ -98,7 +98,7 @@
               <div class="view-advance-request--element">
                 <app-fy-statistic
                   [icon]="'../../../../assets/svg/calendar.svg'"
-                  [type]="'date'"
+                  [type]="statisticType.date"
                   [label]="'Approved On'"
                   [value]="advanceRequest.areq_approved_at | date: 'MMM dd, yyyy'"
                 ></app-fy-statistic>
@@ -110,7 +110,7 @@
         <div class="view-advance-request--element">
           <app-fy-statistic
             [icon]="'../../../../assets/svg/list.svg'"
-            [type]="'string'"
+            [type]="statisticType.string"
             [label]="'Purpose'"
             [value]="advanceRequest.areq_purpose|| '-'"
           ></app-fy-statistic>
@@ -118,9 +118,9 @@
         <div class="view-advance-request--element">
           <app-fy-statistic
             [icon]="'../../../../assets/svg/list.svg'"
-            [type]="'string'"
+            [type]="statisticType.string"
             [label]="'Project'"
-            [value]="advanceRequest.project_name || '-' | ellipsis: 15"
+            [value]="(advanceRequest.project_name || '-') | ellipsis: 15"
           ></app-fy-statistic>
         </div>
 
@@ -129,7 +129,7 @@
         <div class="view-advance-request--element">
           <app-fy-statistic
             [icon]="'../../../../assets/svg/list.svg'"
-            [type]="'string'"
+            [type]="statisticType.string"
             [label]="'Additional Notes'"
             [value]="advanceRequest.areq_notes || '-'"
           ></app-fy-statistic>
@@ -137,45 +137,47 @@
 
         <!-- Custom Fields block -->
 
-        <div *ngIf="(advanceRequestCustomFields$|async)?.length > 0">
-          <div
-            class="view-advance-request--element"
-            *ngFor="let advanceRequestCustomField of advanceRequestCustomFields$ | async; index as i"
-          >
-            <div *ngIf="advanceRequestCustomField.type === 'DATE' && advanceRequestCustomField.value">
-              <app-fy-statistic
-                [icon]="'../../../../assets/svg/calendar.svg'"
-                [type]="'date'"
-                [label]="advanceRequestCustomField.name"
-                [value]="advanceRequestCustomField.value | date: 'MMM dd, yyyy'"
-              ></app-fy-statistic>
-            </div>
-            <div *ngIf="advanceRequestCustomField.type === 'DATE' && !advanceRequestCustomField.value">
-              <app-fy-statistic
-                [icon]="'../../../../assets/svg/calendar.svg'"
-                [type]="'date'"
-                [label]="advanceRequestCustomField.name"
-                [value]="'-'"
-              ></app-fy-statistic>
-            </div>
-            <div *ngIf="advanceRequestCustomField.type === 'BOOLEAN'">
-              <app-fy-statistic
-                [icon]="'../../../../assets/svg/list.svg'"
-                [type]="'boolean'"
-                [label]="advanceRequestCustomField.name"
-                [value]="advanceRequestCustomField.value ? 'Yes' : 'No'"
-              ></app-fy-statistic>
-            </div>
-            <div *ngIf="advanceRequestCustomField.type !== 'BOOLEAN' && advanceRequestCustomField.type !== 'DATE'">
-              <app-fy-statistic
-                [icon]="'../../../../assets/svg/list.svg'"
-                [type]="'boolean'"
-                [label]="advanceRequestCustomField.name"
-                [value]="advanceRequestCustomField.value || '-'"
-              ></app-fy-statistic>
+        <ng-container *ngIf="advanceRequestCustomFields$|async as advanceRequestCustomFields">
+          <div *ngIf="advanceRequestCustomFields.length > 0">
+            <div
+              class="view-advance-request--element"
+              *ngFor="let advanceRequestCustomField of advanceRequestCustomFields; index as i"
+            >
+              <div *ngIf="advanceRequestCustomField.type === 'DATE' && advanceRequestCustomField.value">
+                <app-fy-statistic
+                  [icon]="'../../../../assets/svg/calendar.svg'"
+                  [type]="statisticType.date"
+                  [label]="advanceRequestCustomField.name"
+                  [value]="advanceRequestCustomField.value | date: 'MMM dd, yyyy'"
+                ></app-fy-statistic>
+              </div>
+              <div *ngIf="advanceRequestCustomField.type === 'DATE' && !advanceRequestCustomField.value">
+                <app-fy-statistic
+                  [icon]="'../../../../assets/svg/calendar.svg'"
+                  [type]="statisticType.date"
+                  [label]="advanceRequestCustomField.name"
+                  [value]="'-'"
+                ></app-fy-statistic>
+              </div>
+              <div *ngIf="advanceRequestCustomField.type === 'BOOLEAN'">
+                <app-fy-statistic
+                  [icon]="'../../../../assets/svg/list.svg'"
+                  [type]="statisticType.boolean"
+                  [label]="advanceRequestCustomField.name"
+                  [value]="advanceRequestCustomField.value ? 'Yes' : 'No'"
+                ></app-fy-statistic>
+              </div>
+              <div *ngIf="advanceRequestCustomField.type !== 'BOOLEAN' && advanceRequestCustomField.type !== 'DATE'">
+                <app-fy-statistic
+                  [icon]="'../../../../assets/svg/list.svg'"
+                  [type]="statisticType.boolean"
+                  [label]="advanceRequestCustomField.name"
+                  [value]="advanceRequestCustomField.value || '-'"
+                ></app-fy-statistic>
+              </div>
             </div>
           </div>
-        </div>
+        </ng-container>
 
         <div class="view-advance-request--divider"></div>
       </div>
