+++ src/app/fyle/my-view-advance-request/my-view-advance-request.page.ts
@@ -39,6 +39,13 @@ export class MyViewAdvanceRequestPage implements OnInit {
 
   isDeviceWidthSmall = window.innerWidth < this.minScreenWidth;
 
+  statisticType = {
+    string: 'string',
+    number: 'number',
+    date: 'date',
+    boolean: 'boolean',
+  };
+
   constructor(
     private activatedRoute: ActivatedRoute,
     private loaderService: LoaderService,
